+++ js/spa/test/render_tests.spec.ts
@@ -34,3 +34,18 @@ test("Test event/selection data works in render", async ({ page }) => {
 	await page.getByText("Hi").click();
 	await expect(selected_chat).toHaveValue("[0, 1]");
 });
+
+test("Test event/selection data can trigger render", async ({ page }) => {
+	await page.getByText("chat3").click();
+	const selected_chat = page.getByLabel("Trigger Index");
+	await expect(selected_chat).toHaveValue("[1, 0]");
+});
+
+test("Test examples work in render", async ({ page }) => {
+	await page.getByRole("button", { name: "test" }).click();
+	await expect(page.getByLabel("input", { exact: true })).toHaveValue("test");
+	await page.getByRole("button", { name: "def", exact: true }).click();
+	await expect(page.getByLabel("little textbox", { exact: true })).toHaveValue(
+		"def"
+	);
+});